requirements.txt
anyio==4.8.0
arabic-reshaper==3.0.0
certifi==2024.12.14
convertdate==2.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Khayyam==3.0.17
pillow==11.1.0
PyMeeus==0.5.12
python-bidi==0.6.3
python-telegram-bot==21.10
sniffio==1.3.1
umalqurra==0.2
